CCM-814, CCM-816
Adds mock response for POST /api/v1/send, including the same validation
as the current backend service (at the time of authoring).

Checks include:
* missing request body
* missing sendingGroupId
* missing requestRefId
* missing data array
* missing requestItemRefId within data array
* duplicate requestItemRefId values within data array
* sendingGroupId validation against the following list:
** b838b13c-f98c-4def-93f0-515d4e4f4ee1
** 49e43b98-70cb-47a9-a55e-fe70c9a6f77c
** b402cd20-b62a-4357-8e02-2952959531c8
** 936e9d45-15de-4a95-bb36-ae163c33ae53
** 9ba00d23-cd6f-4aca-8688-00abc85a7980

Includes the ability to simulate a 500 error, which can be thrown
by the existing backend when there is an error writing to the dynamodb
table. This is done by sending a requestRefId value of simulate-500.

Test coverage has been added to excercise each of these eventualities
on the /api/v1/send endpoint.