CCM-820
Adds support for both gateway and request timeouts. On the sandbox
these trigger after 2 seconds.

Adds support for rethrowing 408 and 504 error responses received
from the backend service.

Updated the spec to include the gateway timeout response from
the API.

--

Adds an endpoint on the sandbox app that allows the backend timeout
to be triggered : `GET /_timeout[?sleep=3000]`.

The optional sleep value is set to 3000 milliseconds by default. Passing
an integer value into the call will control the amount of time the
endpoint sleeps for before a response is made.

--

Adds an endpoint for testing the rethrow of 408 error responses from
the backend service : `GET /_timeout_408`.

--

Adds an endpoint for testing the rethrow of 504 error responses from
the backend service : `GET /_timeout_504`.