CCM-956
Adds documentation for the 401 response type, including a link
to the instructions on how to generate a valid bearer token.

Adds the ability to trigger the 401 by sending a request with
the header `Prefer` set to a value of `code=401` - this is to
ensure that the documentation is consistent for end consumers.