CCM-1169
Adds integration test suite with authentication.

Authentication is handled in the same way as API consumers - this is
in contrast to our int-dev/qa environments where we utilise the
NHSD pytest framework when generating credentials.

Tidies up the run-tests.yml pipeline.

Removes unused commands from the Makefile.

Standardises on the names of some files to simplify the pipelines.