CCM-1267
Adds zap security scanner to the nightly test runs. Reports
include both failures and passes. When a failure is detected
the pipeline will fail.

Tidies up the specification to remove invalid server addresses.

Skips the proxy check if on the nightly test run.

Removes run-postman.js script as it is not used.

Swaps over to running the nightly pipeline on the release branch
instead of on master.