CCM-1604
Corrects the 406 response content type so it is set to
application/vnd.api+json.

The reason why it is this, and not application/json is
that our API outputs application/vnd.api+json by default.

On an unacceptable Accept header we must always respond
with the default.

Adds an assertion on the 406 tests to ensure the default
header is returned in the response.