CCM-1559
Renames the repository from `communications-manager` to
`communications-manager-api`.

This is to ensure that there is no mixup between the core
communications manager service and the API which makes it
accessible. As this repository, and the service repository
will be, public - it is important to make this distinction.

Swaps the connection for the azure pipelines over to my
`NHSDigital (63)` integration - we need to raise giving
access to the new github repo with the APIM team to fix
properly.