CCM-1813
Adds sphinx documentation project for test documentation.

Can be run manually via the `make test-documentation` command.

Build step added to ensure failure if the docs have not been
regenerated as part of the commit.

Moved existing documentation under the docs folder and added
a basic index for it.

Adds a basic documentation structure for all test suites.

Adds a placeholder test title in for every test function.

Updates node installation step on the github build action
as the previous method is being deprecated.