CCM-1616 - service bans
Adds a new policy that allows for service ban responses
from the backend solution to trigger a custom 403
response within the API proxy.

Ensures that the sandbox mock application can throw
errors in the same way as the backend application,
with test coverage over it and the policy being run.

Updates the documentation for the proxy to include
the new logic around the handling of 403 & 403 service
bans.

Adds the 403 service ban information to the API
specification.